Se finalizan los test de las operaciones

+++ my-demo/src/main/java/co/com/sofka/app/BasicCalculator.java
@@ -21,7 +21,7 @@ public class BasicCalculator {
         return number1 * number2;
     }
 
-    public Long divide(Long number1, Long number2) {
+    public Double divide(Double number1, Double number2) {
         logger.info( "Dividing {} / {}", number1, number2 );
         if (number2 == 0) {
             throw new IllegalArgumentException("Division by zero");
